clean up store unused api
