+++ legend-depot-core-http/src/main/java/org/finos/legend/depot/core/http/BaseServer.java
@@ -33,7 +33,7 @@ import io.prometheus.client.exporter.MetricsServlet;
 import org.eclipse.jetty.server.session.SessionHandler;
 import org.eclipse.jetty.servlets.CrossOriginFilter;
 import org.eclipse.jetty.util.component.LifeCycle;
-import org.finos.legend.depot.store.api.StorageConfiguration;
+import org.finos.legend.depot.store.StorageConfiguration;
 import org.finos.legend.depot.tracing.configuration.PrometheusMetricsProviderConfiguration;
 import org.finos.legend.depot.tracing.configuration.TracingAuthenticationProviderConfiguration;
 import org.finos.legend.sdlc.server.error.LegendSDLCServerExceptionMapper;
