+++ legend-depot-store-mongo/src/main/java/org/finos/legend/depot/store/mongo/core/MongoClientModule.java
@@ -21,13 +21,13 @@ import com.google.inject.Singleton;
 import com.google.inject.name.Names;
 import com.mongodb.MongoClient;
 import com.mongodb.client.MongoDatabase;
-import org.finos.legend.depot.store.api.StorageConfiguration;
+import org.finos.legend.depot.store.StorageConfiguration;
 import org.finos.legend.depot.tracing.configuration.OpenTracingConfiguration;
 import org.finos.legend.depot.tracing.services.TracerFactory;
 
 import javax.inject.Named;
 
-public class MongoStoreConnectionModule extends PrivateModule
+public class MongoClientModule extends PrivateModule
 {
     @Override
     protected void configure()
