+++ legend-depot-store-server/src/main/java/org/finos/legend/depot/store/guice/DepotStoreResourcesModule.java
@@ -17,7 +17,6 @@ package org.finos.legend.depot.store.guice;
 
 import com.google.inject.PrivateModule;
 import org.finos.legend.depot.store.resources.entities.ManageEntitiesResource;
-import org.finos.legend.depot.store.resources.generation.file.ManageFileGenerationsResource;
 import org.finos.legend.depot.store.resources.projects.ManageProjectsResource;
 import org.finos.legend.depot.store.resources.versions.ManageProjectsVersionsResource;
 
@@ -29,12 +28,10 @@ public class DepotStoreResourcesModule extends PrivateModule
     {
         bind(ManageProjectsResource.class);
         bind(ManageEntitiesResource.class);
-        bind(ManageFileGenerationsResource.class);
         bind(ManageProjectsVersionsResource.class);
 
         expose(ManageProjectsResource.class);
         expose(ManageEntitiesResource.class);
-        expose(ManageFileGenerationsResource.class);
         expose(ManageProjectsVersionsResource.class);
     }
