+++ legend-depot-store-server/src/main/java/org/finos/legend/depot/store/server/configuration/DepotStoreServerConfiguration.java
@@ -19,11 +19,9 @@ import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import org.finos.legend.depot.artifacts.repository.api.ArtifactRepositoryProviderConfiguration;
 import org.finos.legend.depot.core.http.ServersConfiguration;
-import org.finos.legend.depot.domain.project.IncludeProjectPropertiesConfiguration;
+import org.finos.legend.depot.store.artifacts.configuration.IncludeProjectPropertiesConfiguration;
 import org.finos.legend.depot.store.artifacts.configuration.ArtifactsRetentionPolicyConfiguration;
 import org.finos.legend.depot.store.notifications.domain.QueueManagerConfiguration;
-
-import javax.validation.constraints.NotNull;
 
 
 @JsonIgnoreProperties(ignoreUnknown = true)
